requirements.txt
flask
pygithub
requests
pyyaml